+++ src/common/types/character-properties.ts
@@ -93,6 +93,7 @@ export type PassiveTalent = (
   | `While in Nod-Krai, if a member on your team dies, Columbina will revive them and restore health based on Columbina's friendship level. Cooldown 100s. Does not work in domains.`
   | `Collects elemental energy at night`
   | `For every party member from Mondstadt, the Hold version of Varka's Elemental Skill "Windbound Execution"'s cooldown is decreased by 5%.`
+  | `Outside of combat, when Linnea takes an aimed shot, different effects occur depending on target. Specific animals: Arrow transforms into "Net". Ores: Dispatches Lumi who will strike and collect ores.`
 );
 
 export type TriggerableReactionFilter = 'all' | 'playstyle-based';