+++ src/data/materials/drops/elite.ts
@@ -354,3 +354,13 @@ export const XOfTheDeepShadow = MobDrop.create('of the Deep Shadow', {
   [Rarity.Rare]: `The remains of a fisher from beyond this world, left after its body dissolved.\nA core whose functions have fully shut down, and its structure resembles no organ of any creature found on this earth.`,
   [Rarity.Uncommon]: `The remains of a fisher from beyond this world, left after its body dissolved.\nOnce its shell cracked, the malice housed within faded away as well.`
 });
+
+export const FadedFlamingHilt = MobDrop.create('Flaming Hilt', {
+  [Rarity.Epic]: `Jeweled`,
+  [Rarity.Rare]: `Fractured`,
+  [Rarity.Uncommon]: `Faded`,
+}, {
+  [Rarity.Epic]: `The hilt remains of a flaming sword once wielded by a Domain Keeper.\nA sword of blazing fire was once the symbol of admonition.\nEmpowered by the authority of "Space," it proclaimed the absolute and infallible will of the divine.`,
+  [Rarity.Rare]: `The hilt remains of a flaming sword once wielded by a Domain Keeper.\nA sword of blazing fire was once the symbol of judgment.\nEmpowered by the authority of "Space," it barred the path against trespass and profanity.`,
+  [Rarity.Uncommon]: `The hilt remains of a flaming sword once wielded by a Domain Keeper.\nA sword of blazing fire was once the symbol of protection.\nEmpowered by the authority of "Space," it sealed the boundary between eternity and ephemera.`,
+});