+++ change-log.md
@@ -2,7 +2,13 @@
 
 
 ## Genshin Impact v. 6.5
-
+* Added Linnea, 5-star Geo Bow character
+* Added Golden Frostbound Oath, Linnea's signature weapon
+* Added Watcher: Fallen Vigil, Mondstadt World Boss
+* Added Domain Keeper, new Elite mob
+* Added Etherwing Moth, new Mondstadt local specialty
+* Added Plume of the Fallen Watcher, dropped by Watcher: Fallen Vigil
+* Added Faded Flaming Hilt + variations, dropped by Domain Keeper
 
 ## Updates